+++ CHANGELOG.md
@@ -59,6 +59,7 @@ Unless otherwise specified, any version comparison below is the comparison of th
   - Added attack sound when player attacks an entity (`ATTACK_STRONG` on hit, `ATTACK_NODAMAGE` on miss).
   - `PlayerPunchAirEvent` now implements `CancellableEvent`, allowing plugins to cancel the event and prevent the sound from playing.
 - Fixed packet serializer of `PlayerEnchantOptionsPacket` for NetEase 1.21.50.
+- Fixed an issue where too long command outputs will make the client disconnect in 1.21.130+.
 
 # 0.10.6 (API 0.23.0) - 2026/1/24